Fixed NullReferenceException when calling Configure.With().XmlSerializer() prior to assigning a builder, e.g. DefaultBuilder(), AutofacBuilder(), etc.
This also applies to the BinarySerializer.  One other potential way to handle this problem would be to throw a more descriptive exception and
explain the ordering of Configure.With() operations, but the change as committed seemed to be a better way to reduce exception and still get the
desired serializer properly configured.