Do not override EndpointConfig.cs file on updates
Added a check to see if EndpointConfic.cs file already exists to stop overriding it on updates.
Now walking folders to find a IConfigureThisEndpoint implementation.
This patch addresses one of the issues raised in #2057
Unfortunately the other issue about referencing a IConfigureThisEndpoint implementation from a different project or referencing a dll with an IConfigureThisEndpoint implementation is not addressed.