Transports now gets the requested consistency guarantees
Translate old tx config calls to consistency guarantee

Introduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the
messagesIntroduced DispatchConsistency

To allow the outbox to request the message to not enlist in incoming
transaction. This is needed since we mark the outbox message as dispatched
when the .Dispatch doesn't blow. A rollback here would loose the messages