Fix flaky Outbox storage unit test
The root cause of the bug was low resolution of the clock behind
DateTime.UtcNow. DateTime.UtcNow is called inside the test multiple times.
Returned values are later on compared by the < operator to remove old entries
from the outbox storage. In some test runs all the calls to DateTime.UtcNow
return the same value and the < operator on these values cannot be used
reliably to remove old entries. The simplest solution is to artificially "delay"
the moment we capture storage time by 1 CPU tick.

Another, less elegant, option would be to introduce artificial delay in the form
of Thread.Sleep or Task.Delay, but adding 1 tick is more reliable.