removing no longer needed test
I think we added it just as a safeguard while we were testing the first solutions. To detect if we mess up with this scenario. We didn't add any extra routing state in the behavior send because the other scenario (the original that uses an outgoing scenario) didn't need it, it failed just by using SendLocal from the endpoint. Removing the test for now.